Solving Responsiveness Issues
Previously, the first row of buttons of the calculator would look misaligned when the screen width became low.
Now, the width needs to decrease to an unreasonable level for the issue to recur.
This was done by changing the text of the DEL button to a left facing arrow and that of the AC button to just C.

Limited number of decimal places visible for division and multiplication to 5.